backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
python-dotenv
google-auth
google-auth-oauthlib
google-api-python-client
gspread
google-generativeai
google-genai
httpx
